feat: delete 챗룸 소유주인지 확인

+++ src/chat/chat.service.ts
@@ -383,7 +383,14 @@ export class ChatService {
 	 * @param roomId
 	 * @returns
 	 */
-	deleteChatRoom(roomId: number): Promise<void> {
+	async deleteChatRoom(roomId: number, userId: number): Promise<void> {
+		const participant = await this.chatParticipantRepository.getParticipant(
+			roomId,
+			userId
+		);
+		if (!participant || participant.role !== Role.OWNER) {
+			throw new UnauthorizedException();
+		}
 		return this.chatRoomRepository.deleteChatRoom(roomId);
 	}
